Implement InlinedGlobalMarkingVisitor
This CL introduces inlined version of MarkingVisitor<GlobalMarking>.
InlinedGlobalMarkingVisitor emulates an Visitor*, but all time critial
methods are inlined.

This CL doesn't yet enable use of InlinedGlobalMarkingVisitor.
InlinedGlobalMarkingVisitor is to be enabled in future CLs.

Most of the existing implementation of MarkingVisitor is
moved to MarkingVisitorImpl.h, so that implementation is
shared between InlinedGlobalMarkingVisitor and MarkingVisitor<>.

BUG=420515

Review URL: https://codereview.chromium.org/816563002

git-svn-id: svn://svn.chromium.org/blink/trunk@187521 bbb929c8-8fbe-4397-9dbb-9b2b20218538