Demistify DocumentParser::append
Today it takes a PassRefptr<StringImpl> as argument. This supposed to express
that the input string is passed in a destructive manner. However this is no
longer the case. It was needed by the background parser but this code has been
refactored and now append is only used on the synchronous parsing path.

Also remove outdated comment that it should not be public. Both DocumentWriter
and DOMPatchSupport uses it and there is nothing particularly wrong with that.

R=kouhei@chromium.org

Review URL: https://codereview.chromium.org/799803004

git-svn-id: svn://svn.chromium.org/blink/trunk@187561 bbb929c8-8fbe-4397-9dbb-9b2b20218538