Remove DOMWindowLifecycleObserver::window()
This is a preparation for landing https://codereview.chromium.org/822863003/. I'm planning to make LocalDOMWindow's destruction unobservable because the timing becomes undeterministic in a GC system. This CL removes DOMWindowLifecycleObserver::window(), which is a method whose behavior depends on a timing of LocalDOMWindow's destruction.

BUG=340522

Review URL: https://codereview.chromium.org/826223002

git-svn-id: svn://svn.chromium.org/blink/trunk@187750 bbb929c8-8fbe-4397-9dbb-9b2b20218538